backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
rembg==2.0.57
Pillow==10.4.0
numpy==1.26.0
